Revert test changes made for jlinkOptions
