+++ src/main/java/org/openjfx/JavaFXBaseMojo.java
@@ -61,7 +61,6 @@ import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Properties;
-import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
@@ -445,42 +444,6 @@ abstract class JavaFXBaseMojo extends AbstractMojo {
             return null;
         }
         return path.getRoot().resolve(path.subpath(0, path.getNameCount() - depth));
-    }
-
-    List<String> splitComplexArgumentString(String argumentString) {
-        char[] strArr = argumentString.trim().toCharArray();
-
-        List<String> splitedArgs = new ArrayList<>();
-        StringBuilder sb = new StringBuilder();
-
-        char expectedSeparator = ' ';
-        for (int i = 0; i < strArr.length; i++) {
-            char item = strArr[i];
-
-            if (item == expectedSeparator
-                    || (expectedSeparator == ' ' && Pattern.matches("\\s", String.valueOf(item))) ) {
-
-                if (expectedSeparator == '"' || expectedSeparator == '\'') {
-                    sb.append(item);
-                    expectedSeparator = ' ';
-                } else if (expectedSeparator == ' ' && sb.length() > 0) {
-                    splitedArgs.add(sb.toString());
-                    sb.delete(0, sb.length());
-                }
-            } else {
-                if (expectedSeparator == ' ' && (item == '"' || item == '\'')) {
-                    expectedSeparator = item;
-                }
-
-                sb.append(item);
-            }
-
-            if (i == strArr.length - 1 && sb.length() > 0) {
-                splitedArgs.add(sb.toString());
-            }
-        }
-
-        return splitedArgs;
     }
 
     private static String findExecutable(final String executable, final List<String> paths) {
