+++ src/main/java/org/openjfx/JavaFXRunMojo.java
@@ -38,6 +38,7 @@ import java.util.Collection;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
+import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 
 import static org.openjfx.model.RuntimePathOption.CLASSPATH;
@@ -185,6 +186,42 @@ public class JavaFXRunMojo extends JavaFXBaseMojo {
             return moduleDescriptor.name() + "/" + mainClass;
         }
         return mainClass;
+    }
+
+    private List<String> splitComplexArgumentString(String argumentString) {
+        char[] strArr = argumentString.trim().toCharArray();
+
+        List<String> splitedArgs = new ArrayList<>();
+        StringBuilder sb = new StringBuilder();
+
+        char expectedSeparator = ' ';
+        for (int i = 0; i < strArr.length; i++) {
+            char item = strArr[i];
+
+            if (item == expectedSeparator
+                    || (expectedSeparator == ' ' && Pattern.matches("\\s", String.valueOf(item))) ) {
+
+                if (expectedSeparator == '"' || expectedSeparator == '\'') {
+                    sb.append(item);
+                    expectedSeparator = ' ';
+                } else if (expectedSeparator == ' ' && sb.length() > 0) {
+                    splitedArgs.add(sb.toString());
+                    sb.delete(0, sb.length());
+                }
+            } else {
+                if (expectedSeparator == ' ' && (item == '"' || item == '\'')) {
+                    expectedSeparator = item;
+                }
+
+                sb.append(item);
+            }
+
+            if (i == strArr.length - 1 && sb.length() > 0) {
+                splitedArgs.add(sb.toString());
+            }
+        }
+
+        return splitedArgs;
     }
 
     // for tests
